move status callout to the side panel
