+++ assets/src/components/ai/agent-runs/details/AgentRunWorkingTheory.tsx
@@ -1,0 +1,260 @@
+import {
+  ArrowTopRightIcon,
+  Button,
+  Card,
+  Flex,
+  IconFrame,
+  PrIcon,
+  Toast,
+} from '@pluralsh/design-system'
+import { AgentTodosTimeline } from 'components/ai/common/AgentTodosTimeline'
+import { WorkbenchLinkChip } from 'components/workbenches/common/WorkbenchLinkChip'
+import { StretchedFlex } from 'components/utils/StretchedFlex'
+import { StackedText } from 'components/utils/table/StackedText'
+import { Body2P, Subtitle1H1 } from 'components/utils/typography/Text'
+import {
+  AgentRunFragment,
+  AgentRunStatus,
+  AgentTodoFragment,
+  useApproveAgentRunMutation,
+} from 'generated/graphql'
+import { useTheme } from 'styled-components'
+import { isNonNullable } from 'utils/isNonNullable'
+
+function normalizeTodos(todos: AgentTodoFragment[]) {
+  return todos
+    .filter(isNonNullable)
+    .map((todo) => ({
+      ...todo,
+      title: todo.title?.trim() ?? '',
+      description: todo.description?.trim() ?? '',
+      done: !!todo.done,
+    }))
+    .filter((todo) => todo.title.length > 0 || todo.description.length > 0)
+}
+
+export function AgentRunWorkingTheory({
+  run,
+  todos: rawTodos,
+  onViewDiff,
+}: {
+  run: AgentRunFragment
+  todos: AgentTodoFragment[]
+  onViewDiff: () => void
+}) {
+  const [approveAgentRun, { loading: approving, error: approvingError }] =
+    useApproveAgentRunMutation({
+      variables: { id: run.id },
+      refetchQueries: ['AgentRun', 'PendingApprovalAgentRuns'],
+    })
+
+  const isApprovable =
+    run.status === AgentRunStatus.PendingApproval && !run.approvedAt
+
+  const showStatusCallout = shouldShowAgentRunStatusCallout(run)
+
+  const todos = normalizeTodos(rawTodos)
+
+  return (
+    <Flex
+      direction="column"
+      gap="xlarge"
+    >
+      {showStatusCallout && (
+        <AgentRunStatusCallout
+          run={run}
+          isApprovable={isApprovable}
+          approving={approving}
+          onApprove={() => approveAgentRun()}
+          onViewDiff={onViewDiff}
+        />
+      )}
+      <Subtitle1H1>Agent todos</Subtitle1H1>
+      <AgentTodosTimeline todos={todos} />
+      <Toast
+        error="Approving agent run failed"
+        show={!!approvingError}
+        closeTimeout={5000}
+        severity="danger"
+        position="bottom"
+        marginBottom="medium"
+      >
+        {approvingError?.message}
+      </Toast>
+    </Flex>
+  )
+}
+
+export function shouldShowAgentRunStatusCallout(
+  run: Nullable<AgentRunFragment>
+) {
+  return (
+    !!run &&
+    ((run.pullRequests ?? []).some(Boolean) ||
+      !!run.analysis?.summary ||
+      (run.status === AgentRunStatus.PendingApproval && !run.approvedAt))
+  )
+}
+
+function AgentRunStatusCallout({
+  run,
+  isApprovable,
+  approving,
+  onApprove,
+  onViewDiff,
+}: {
+  run: AgentRunFragment
+  isApprovable: boolean
+  approving: boolean
+  onApprove: () => void
+  onViewDiff: () => void
+}) {
+  const theme = useTheme()
+  const pullRequest = run.pullRequests?.[0]
+  const title = pullRequest?.title ?? agentRunStatusTitle(run.status)
+  const summary = run.analysis?.summary
+  const hasPatch = !!run.upload?.patch
+  const workbenchJob = run.workbenchJob
+  const workbench = workbenchJob?.workbench
+  const showWorkbenchChip =
+    !!workbenchJob?.id && !!workbench?.id && !!workbench.name
+
+  return (
+    <Card
+      fillLevel={1}
+      css={{
+        display: 'flex',
+        flexDirection: 'column',
+        gap: theme.spacing.small,
+        padding: theme.spacing.medium,
+        width: '100%',
+        borderLeft: `3px solid ${theme.colors[statusToBorderColor[run.status]]}`,
+      }}
+    >
+      <StretchedFlex
+        align="start"
+        gap="medium"
+      >
+        <StackedText
+          truncate
+          first={title}
+          firstPartialType="body2Bold"
+          firstColor="text-light"
+          second={pullRequest?.title ? agentRunStatusTitle(run.status) : null}
+          secondColor="text-xlight"
+          icon={
+            <IconFrame
+              circle
+              size="large"
+              type="secondary"
+              icon={
+                <PrIcon
+                  size="small"
+                  color="icon-light"
+                />
+              }
+              css={{ flexShrink: 0 }}
+            />
+          }
+          css={{ flex: 1, minWidth: 0 }}
+        />
+      </StretchedFlex>
+      {summary && (
+        <Body2P
+          $color="text-light"
+          css={{
+            display: '-webkit-box',
+            WebkitLineClamp: 3,
+            WebkitBoxOrient: 'vertical',
+            overflow: 'hidden',
+          }}
+        >
+          {summary}
+        </Body2P>
+      )}
+      {(hasPatch || pullRequest?.url || isApprovable || showWorkbenchChip) && (
+        <StretchedFlex
+          align="center"
+          gap="small"
+        >
+          {showWorkbenchChip && (
+            <WorkbenchLinkChip
+              workbenchId={workbench.id}
+              workbenchName={workbench.name}
+              workbenchJobId={workbenchJob.id}
+              css={{ flexShrink: 0 }}
+            />
+          )}
+          {(hasPatch || pullRequest?.url || isApprovable) && (
+            <Flex
+              gap="small"
+              css={{ marginLeft: 'auto' }}
+            >
+              {hasPatch ? (
+                <Button
+                  small
+                  secondary
+                  onClick={onViewDiff}
+                >
+                  View diff
+                </Button>
+              ) : (
+                pullRequest?.url && (
+                  <Button
+                    small
+                    secondary
+                    as="a"
+                    href={pullRequest.url}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    endIcon={<ArrowTopRightIcon size={12} />}
+                  >
+                    View PR
+                  </Button>
+                )
+              )}
+              {isApprovable && (
+                <Button
+                  small
+                  onClick={onApprove}
+                  loading={approving}
+                >
+                  Approve & create PR
+                </Button>
+              )}
+            </Flex>
+          )}
+        </StretchedFlex>
+      )}
+    </Card>
+  )
+}
+
+function agentRunStatusTitle(status: AgentRunStatus) {
+  switch (status) {
+    case AgentRunStatus.PendingApproval:
+      return 'Approval required'
+    case AgentRunStatus.Successful:
+      return 'Run complete'
+    case AgentRunStatus.Failed:
+      return 'Run failed'
+    case AgentRunStatus.Cancelled:
+      return 'Run cancelled'
+    case AgentRunStatus.Babysitting:
+      return 'Babysitting'
+    case AgentRunStatus.Running:
+      return 'Agent run in progress'
+    case AgentRunStatus.Pending:
+      return 'Agent run pending'
+  }
+}
+
+const statusToBorderColor = {
+  [AgentRunStatus.PendingApproval]: 'icon-warning',
+  [AgentRunStatus.Successful]: 'icon-success',
+  [AgentRunStatus.Failed]: 'icon-danger',
+  [AgentRunStatus.Cancelled]: 'icon-xlight',
+  [AgentRunStatus.Babysitting]: 'icon-info',
+  [AgentRunStatus.Running]: 'icon-info',
+  [AgentRunStatus.Pending]: 'icon-info',
+} as const
